Plugins: Add support for custom node printers
