+++ .changeset/sharp-plants-remain.md
@@ -1,0 +1,5 @@
+---
+'esrap': minor
+---
+
+Add support for custom node printers
