+++ src/types.d.ts
@@ -2,10 +2,6 @@ import { TSESTree } from '@typescript-eslint/types';
 
 type Handler<T> = (node: T, state: State) => undefined;
 
-export type Handlers = {
-	[T in TSESTree.Node['type']]: Handler<Extract<TSESTree.Node, { type: T }>>;
-};
-
 export type TypeAnnotationNodes =
 	| TSESTree.TypeNode
 	| TSESTree.TypeElement
@@ -18,6 +14,10 @@ export type TypeAnnotationNodes =
 	| TSESTree.TSInterfaceHeritage
 	| TSESTree.TSClassImplements
 	| TSExpressionWithTypeArguments;
+	
+export type Handlers = {
+	[T in TSESTree.Node['type']]?: Handler<Extract<TSESTree.Node, { type: T }>>;
+};
 
 type TSExpressionWithTypeArguments = {
 	type: 'TSExpressionWithTypeArguments';
@@ -35,6 +35,7 @@ export interface State {
 	commands: Command[];
 	comments: TSESTree.Comment[];
 	multiline: boolean;
+	handlers: Handlers;
 	quote: "'" | '"';
 }
 
@@ -74,4 +75,5 @@ export interface PrintOptions {
 	sourceMapEncodeMappings?: boolean; // default true
 	indent?: string; // default tab
 	quotes?: 'single' | 'double'; // default single
+	handlers?: Handlers; // default to ...js, ...ts
 }
